refactor(Host): RHICOMPL-1138 introduce HostSearching module
One more filter will cause the Host model to exceed rubocop's class
length cop. We are adding a new filter term to search for hosts tested
against a given policy (or profile) ID.
